requirements.txt
django==5.2
djangorestframework==3.14
neomodel==5.1
neo4j-driver==5.8
load_dotenv